refactor(dashboard): extract lifecycle domain from keeper.ts
Split boot/shutdown/reset/clear/checkpoints/pause/resume/wake/bulk into
keeper-lifecycle.ts (307 lines). keeper.ts ->      873 lines.

Self-contained domain: isRecord + core (fetchWithTimeout/jsonHeaders/
DEFAULT_GET/POST_TIMEOUT_MS) deps; safeJsonResponse, safeKeeperLifecycle,
KeeperLifecycleResponse, KeeperClearRequest stay internal (non-exported).
KEEPER_LIFECYCLE_TIMEOUT_MS import dropped from keeper.ts (only lifecycle
used it). Facade re-export preserves 15 public symbols; consumers unchanged.

typecheck green; 4 test files / 31 tests pass.
